don't install paru if existed

+++ install_scripts/install_aur_helper.sh
@@ -1,9 +1,12 @@
 #!/bin/bash
 set -euox pipefail
 
-sudo pacman -S --needed base-devel
-git clone https://aur.archlinux.org/paru-bin.git
-cd $(pwd)/paru-bin
-makepkg -si --noconfirm
-cd ../
-rm $(pwd)/paru-bin -rf
+if [ -z $(which paru) ]
+then
+    sudo pacman -S --needed git base base-devel
+    git clone https://aur.archlinux.org/paru-bin.git
+    cd $(pwd)/paru-bin
+    makepkg -si --noconfirm
+    cd ../
+    rm $(pwd)/paru-bin -rf
+fi
